Move import resolution task into the `task` package
