+++ tools/mc-js/src/test/java/io/spine/tools/mc/js/code/task/ResolveImportsTest.java
@@ -24,14 +24,14 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.tools.mc.js.code.imports;
+package io.spine.tools.mc.js.code.task;
 
 import com.google.common.collect.ImmutableList;
 import com.google.common.truth.IterableSubject;
-import io.spine.tools.js.fs.Directory;
-import io.spine.tools.mc.js.code.given.GivenProject;
 import io.spine.tools.fs.DirectoryPattern;
 import io.spine.tools.fs.ExternalModule;
+import io.spine.tools.js.fs.Directory;
+import io.spine.tools.mc.js.code.given.GivenProject;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
